requirements.txt
# Core Dependencies
mysql-connector-python==8.2.0
pandas==2.1.4
numpy==1.26.2

# Testing
pytest==7.4.3
pytest-cov==4.1.0

# Performance and Utilities
python-dotenv==1.0.0
colorama==0.4.6

# Optional: For visualization and reporting
matplotlib==3.8.2
seaborn==0.13.0
tabulate==0.9.0
